[JENKINS-13995]
Added logging more liberally, and fixed the problem.
The root cause was that when the mouse enters a hot spot, I wasn't cancelling the previous delayed hide timer.

This fix eliminates the need for the hack in the hide method where I was testing if the mouse was still over the hotspot.

Conflicts:
	changelog.html

Signed-off-by: Oleg Nenashev <nenashev@synopsys.com>